pytest: Separate user tests
