+++ server/camcops_server/cc_modules/cc_user.py
@@ -64,7 +64,6 @@ from camcops_server.cc_modules.cc_sqla_coltypes import (
 )
 from camcops_server.cc_modules.cc_sqlalchemy import Base
 from camcops_server.cc_modules.cc_text import TERMS_CONDITIONS_UPDATE_DATE
-from camcops_server.cc_modules.cc_unittest import DemoDatabaseTestCase
 
 if TYPE_CHECKING:
     from camcops_server.cc_modules.cc_patient import Patient
@@ -1190,66 +1189,3 @@ def set_password_directly(req: "CamcopsRequest",
     user.enable(req)
     audit(req, "Password changed for user " + user.username, from_console=True)
     return True
-
-
-# =============================================================================
-# Unit testing
-# =============================================================================
-
-class UserTests(DemoDatabaseTestCase):
-    """
-    Unit tests.
-    """
-    def test_user(self) -> None:
-        self.announce("test_user")
-        req = self.req
-
-        SecurityAccountLockout.delete_old_account_lockouts(req)
-        self.assertIsInstance(
-            SecurityAccountLockout.is_user_locked_out(req, "dummy_user"),
-            bool
-        )
-        self.assertIsInstanceOrNone(
-            SecurityAccountLockout.user_locked_out_until(req, "dummy_user"),
-            Pendulum
-        )
-
-        self.assertIsInstance(
-            SecurityLoginFailure.how_many_login_failures(req, "dummy_user"),
-            int
-        )
-        SecurityLoginFailure.clear_login_failures_for_nonexistent_users(req)
-        SecurityLoginFailure.clear_dummy_login_failures_if_necessary(req)
-        SecurityLoginFailure.clear_dummy_login_failures_if_necessary(req)
-        # ... do it twice (we had a bug relating to offset-aware vs
-        # offset-naive date/time objects).
-
-        self.assertIsInstance(User.is_username_permissible("some_user"), bool)
-        User.take_some_time_mimicking_password_encryption()
-
-        u = self.dbsession.query(User).first()  # type: User
-        assert u, "Missing user in demo database!"
-
-        g = self.dbsession.query(Group).first()  # type: Group
-        assert g, "Missing group in demo database!"
-
-        self.assertIsInstance(u.is_password_valid("dummy_password"), bool)
-        self.assertIsInstance(u.must_agree_terms, bool)
-        u.agree_terms(req)
-        u.clear_login_failures(req)
-        self.assertIsInstance(u.is_locked_out(req), bool)
-        self.assertIsInstanceOrNone(u.locked_out_until(req), Pendulum)
-        u.enable(req)
-        self.assertIsInstance(u.may_login_as_tablet, bool)
-        # TODO: cc_user.UserTests: could do more here
-        self.assertIsInstance(u.authorized_as_groupadmin, bool)
-        self.assertIsInstance(u.may_use_webviewer, bool)
-        self.assertIsInstance(u.authorized_to_add_special_note(g.id), bool)
-        self.assertIsInstance(u.authorized_to_erase_tasks(g.id), bool)
-        self.assertIsInstance(u.authorized_to_dump, bool)
-        self.assertIsInstance(u.authorized_for_reports, bool)
-        self.assertIsInstance(u.may_view_all_patients_when_unfiltered, bool)
-        self.assertIsInstance(u.may_view_no_patients_when_unfiltered, bool)
-        self.assertIsInstance(u.may_upload_to_group(g.id), bool)
-        self.assertIsInstance(u.may_upload, bool)
-        self.assertIsInstance(u.may_register_devices, bool)
